+++ ricaun.Revit.UI.Example/Models/TestModel.cs
@@ -6,7 +6,7 @@ namespace ricaun.Revit.UI.Example.Models
     public class TestModel : NotifyPropertyBase
     {
         public string Text { get; set; }
-
         public ICommand CommandTest { get; set; }
+        public ICommand CommandTest2 { get; set; }
     }
 }