+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.0.8] / 2022-02-14
+- Move to Button/Pulldown/Split/ComboBox Extensions
+- Move VerifyNameExclusive to RibbonSafeExtension
 - Update RelayCommand
 - 2022-02-12
 - Example Add RelayCommand and Test Some Things to Detect Revit Busy